# Tuan01/github-copilot-vibe-coding-workshop/complete/python/openapi_from_app.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Simple Social Media API",
    "description": "A basic Social Networking Service (SNS) API that allows users to create, retrieve, update, and delete posts; add comments; and like/unlike posts.",
    "contact": {
      "name": "Contoso Product Team",
      "email": "support@contoso.com"
    },
    "license": { "name": "MIT", "url": "https://opensource.org/licenses/MIT" },
    "version": "1.0.0"
  },
  "paths": {
    "/openapi.json": {
      "get": {
        "summary": "Get Openapi",
        "description": "Return the OpenAPI specification.",
        "operationId": "get_openapi_openapi_json_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": {} } }
          }
        }
      }
    },
    "/api/posts": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get Posts",
        "description": "List all posts - Retrieve all recent posts to browse what others are sharing.",
        "operationId": "get_posts_api_posts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": { "$ref": "#/components/schemas/Post" },
                  "type": "array",
                  "title": "Response Get Posts Api Posts Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Posts"],
        "summary": "Create New Post",
        "description": "Create a new post - Create a new post to share something with others.",
        "operationId": "create_new_post_api_posts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewPostRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}": {
      "get": {
        "tags": ["Posts"],
        "summary": "Get Post By Id Endpoint",
        "description": "Get a specific post - Retrieve a specific post by its ID to read in detail.",
        "operationId": "get_post_by_id_endpoint_api_posts__post_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Posts"],
        "summary": "Update Post Endpoint",
        "description": "Update a post - Update an existing post if you made a mistake or have something to add.",
        "operationId": "update_post_endpoint_api_posts__post_id__patch",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdatePostRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Post" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Posts"],
        "summary": "Delete Post Endpoint",
        "description": "Delete a post - Delete a post if you no longer want it shared.",
        "operationId": "delete_post_endpoint_api_posts__post_id__delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}/comments": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get Comments By Post Id Endpoint",
        "description": "List comments for a post - Retrieve all comments on a specific post.",
        "operationId": "get_comments_by_post_id_endpoint_api_posts__post_id__comments_get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Comment" },
                  "title": "Response Get Comments By Post Id Endpoint Api Posts  Post Id  Comments Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Comments"],
        "summary": "Create Comment Endpoint",
        "description": "Create a comment - Add a comment to a post to share your thoughts.",
        "operationId": "create_comment_endpoint_api_posts__post_id__comments_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewCommentRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Comment" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}/comments/{comment_id}": {
      "get": {
        "tags": ["Comments"],
        "summary": "Get Comment By Id Endpoint",
        "description": "Get a specific comment - Retrieve a specific comment by its ID.",
        "operationId": "get_comment_by_id_endpoint_api_posts__post_id__comments__comment_id__get",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Comment Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Comment" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Comments"],
        "summary": "Update Comment Endpoint",
        "description": "Update a comment - Update an existing comment to correct or revise it.",
        "operationId": "update_comment_endpoint_api_posts__post_id__comments__comment_id__patch",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Comment Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCommentRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Comment" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Comments"],
        "summary": "Delete Comment Endpoint",
        "description": "Delete a comment - Delete a comment if necessary.",
        "operationId": "delete_comment_endpoint_api_posts__post_id__comments__comment_id__delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          },
          {
            "name": "comment_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Comment Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/posts/{post_id}/likes": {
      "post": {
        "tags": ["Likes"],
        "summary": "Like Post Endpoint",
        "description": "Like a post - Like a post to show appreciation.",
        "operationId": "like_post_endpoint_api_posts__post_id__likes_post",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LikeRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LikeResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Likes"],
        "summary": "Unlike Post Endpoint",
        "description": "Unlike a post - Remove your like from a post if you change your mind.",
        "operationId": "unlike_post_endpoint_api_posts__post_id__likes_delete",
        "parameters": [
          {
            "name": "post_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Post Id" }
          },
          {
            "name": "username",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Username of the user removing the like",
              "title": "Username"
            },
            "description": "Username of the user removing the like"
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Comment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the comment"
          },
          "postId": {
            "type": "string",
            "title": "Postid",
            "description": "ID of the post this comment belongs to"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username of the comment author"
          },
          "content": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content",
            "description": "Content of the comment"
          },
          "createdAt": {
            "type": "string",
            "title": "Createdat",
            "description": "Timestamp when the comment was created"
          },
          "updatedAt": {
            "type": "string",
            "title": "Updatedat",
            "description": "Timestamp when the comment was last updated"
          }
        },
        "type": "object",
        "required": [
          "id",
          "postId",
          "username",
          "content",
          "createdAt",
          "updatedAt"
        ],
        "title": "Comment"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LikeRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Username of the user liking the post"
          }
        },
        "type": "object",
        "required": ["username"],
        "title": "LikeRequest"
      },
      "LikeResponse": {
        "properties": {
          "postId": {
            "type": "string",
            "title": "Postid",
            "description": "ID of the liked post"
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Username of the user who liked the post"
          },
          "likedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Likedat",
            "description": "Timestamp when the post was liked"
          }
        },
        "type": "object",
        "required": ["postId", "username", "likedAt"],
        "title": "LikeResponse"
      },
      "NewCommentRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username of the comment author"
          },
          "content": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content",
            "description": "Content of the comment"
          }
        },
        "type": "object",
        "required": ["username", "content"],
        "title": "NewCommentRequest"
      },
      "NewPostRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username of the post author"
          },
          "content": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "description": "Content of the post"
          }
        },
        "type": "object",
        "required": ["username", "content"],
        "title": "NewPostRequest"
      },
      "Post": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique identifier for the post"
          },
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username of the post author"
          },
          "content": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "description": "Content of the post"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat",
            "description": "Timestamp when the post was created"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat",
            "description": "Timestamp when the post was last updated"
          },
          "likesCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Likescount",
            "description": "Number of likes on the post"
          },
          "commentsCount": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Commentscount",
            "description": "Number of comments on the post"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "content",
          "createdAt",
          "updatedAt",
          "likesCount",
          "commentsCount"
        ],
        "title": "Post"
      },
      "UpdateCommentRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username of the comment author (for validation)"
          },
          "content": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Content",
            "description": "Updated content of the comment"
          }
        },
        "type": "object",
        "required": ["username", "content"],
        "title": "UpdateCommentRequest"
      },
      "UpdatePostRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "description": "Username of the post author (for validation)"
          },
          "content": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Content",
            "description": "Updated content of the post"
          }
        },
        "type": "object",
        "required": ["username", "content"],
        "title": "UpdatePostRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}